Create a seperate sort control component
